+++ sample_data/file007_truncated.txt
