+++ scripts/interface/cli.py
@@ -1,0 +1,1 @@
+print
