+++ scripts/plotting/accel/acceleration_change_grapher.py
@@ -1,16 +1,15 @@
-""" A function used to graph acceleration from a file of sheep accelerometer data, using numpy and matplotlib in python.
+""" A function used to graph change in acceleration from a file of sheep accelerometer data, using numpy and matplotlib in 
+python.
 Assumptions: accelerometer records data at 25Hz, files are formatted as a line of timestamps and 25 lines of accelerometer 
-data after each timestamp.
+data after each timestamp, timestamp is formatted as (*seconds, milliseconds).
 """
 
+import matplotlib
 import matplotlib.pyplot as plt
 import numpy as np
 import math
 
 FREQUENCY = 25
-# from scipy import constants
-# GRAVITY = constants.g
-GRAVITY   = 9.80665
 
 def int_list(data):
     '''Converts elements of a list to type int'''
@@ -18,7 +17,7 @@ def int_list(data):
         data[i] = int(data[i])
 
 def float_list(data):
-    'Converts elements of a list to type float'
+    '''Converts elements of a list to type float'''
     for i in range(len(data)):
         data[i] = float(data[i])
 
@@ -44,25 +43,20 @@ def timestamp_to_seconds(times):
     '''Returns time in milliseconds, given a numpy array of seconds and milliseconds'''
     return 1000 * times[0] + times[1]
 
-def acceleration_grapher(filename, header = False, calibrate_x = [-1.0,1.0], calibrate_y = [-1.0,1.0], calibrate_z = [-1.0,1.0]):
-    """Returns an acceleration line graph pyplot object, given a file of sheep accelerometer data
-    
-    @param: calibrate_x a list of size 2 with a calibration lower bound and upper bound 
-    @param: calibrate_y a list of size 2 with a calibration lower bound and upper bound 
-    @param: calibrate_z a list of size 2 with a calibration lower bound and upper bound 
-    @return: a pyplot object containing the acceleration line graph
-    """
-    x_acceleration   = np.array([0])
-    y_acceleration   = np.array([0])
-    z_acceleration   = np.array([0])
+def acceleration_change_grapher(filename, header = False):
+    '''Returns an acceleration change line graph pyplot object, given a file of sheep accelerometer data'''
+    acceleration_y_axis = np.array([0])
     time_x_axis = np.array([0])
     with open(filename) as file:
         lines = file.readlines()
         if not header:
-            previous_timestamp = extract_timestamp(lines[0])
+            previous_timestamp    = extract_timestamp(lines[0])
+            previous_acceleration = lines[1].strip().split(',')
         else:
-            previous_timestamp = extract_timestamp(lines[1])
+            previous_timestamp    = extract_timestamp(lines[1])
+            previous_acceleration = lines[2].strip().split(',')
         initial_timestamp = previous_timestamp
+        float_list(previous_acceleration)
         for line in lines:
             # Checks for timestamp which starts with *
             if line.startswith('*'):
@@ -77,22 +71,15 @@ def acceleration_grapher(filename, header = False, calibrate_x = [-1.0,1.0], cal
                     zero_line_count += 1
                 acceleration = line[:3].strip().split(',')
                 float_list(acceleration) 
-                x_acceleration   = np.append(acceleration[0])
-                y_acceleration   = np.append(acceleration[1])
-                z_acceleration   = np.append(acceleration[2])
+                acceleration_y_axis = np.append(acceleration_y_axis, euclidean_distance(previous_acceleration, acceleration))
+                previous_acceleration = acceleration
     extend_x_axis = np.linspace(timestamp_to_seconds(timestamp), timestamp_to_seconds(timestamp) + 1, FREQUENCY - zero_line_count)
     time_x_axis = np.append(time_x_axis, np.delete(extend_x_axis, 0))
-    # Calibrates x,y,z accelerometer units into gravity units
-    x_acceleration = (2 * x_acceleration - calibrate_x[0] - calibrate_x[1]) / (calibrate_x[1] - calibrate_x[0])
-    y_acceleration = (2 * y_acceleration - calibrate_y[0] - calibrate_y[1]) / (calibrate_y[1] - calibrate_y[0])
-    z_acceleration = (2 * z_acceleration - calibrate_z[0] - calibrate_z[1]) / -(calibrate_z[1] - calibrate_z[0])
-    net_acceleration = np.sqrt(x_acceleration ** 2 + y_acceleration ** 2 + z_acceleration ** 2)
-    plt.plot(time_x_axis, net_acceleration)
-    plt.title(f"Acceleration for {filename}")
+    plt.plot(time_x_axis, acceleration_y_axis)
+    plt.title(f"Change in acceleration for {filename}")
     plt.xlabel("Timestamp (s)")
-    plt.ylabel("Acceleration (g)")
-    # Y-axis label for uncalibrated data
-    if calibrate_x == [-1.0,1.0] or calibrate_y == [-1.0,1.0] or calibrate_z == [-1.0,1.0]:
-        plt.ylabel("Acceleration (units/s^2)")
+    plt.ylabel("Acceleration (units/s^2)")
     plt.show()
     return plt
+
+acceleration_change_grapher('resources/file007.txt')
