+++ scripts/plotting/fft/analysis.Rmd
