+++ scripts/plotting/fft/process_data.py
