+++ scripts/plotting/freq/plot_frequency2.py
