+++ scripts/plotting/freq/xFrequencypersec_files/MathJax.js
