+++ start.py
@@ -1,0 +1,3 @@
+print("Hello!")
+
+# ask user if user wants CLI or GUI
